+++ src/main/java/com/schnabel/schnabel/drugs/service/DrugService.java
@@ -10,7 +10,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 @Service
-public class DrugService extends CrudService<Drug, Integer> implements IDrugService
+public class DrugService extends CrudService<Drug, Long> implements IDrugService
 {
 
     @Autowired
