+++ src/main/java/com/schnabel/schnabel/drugs/service/IDrugService.java
@@ -2,7 +2,7 @@ package com.schnabel.schnabel.drugs.service;
 import com.schnabel.schnabel.drugs.model.Drug;
 import com.schnabel.schnabel.misc.interfaces.ICrudService;
 
-public interface IDrugService extends ICrudService<Drug, Integer>
+public interface IDrugService extends ICrudService<Drug, Long>
 {
     boolean getByName(String name);
 }
